Refactor EventListCard as a shared component (#6334)
