+++ src/shared-components/EventListCard/EventListCard.tsx
