+++ src/shared-components/EventListCard/Modal/Delete/EventListCardDeleteModal.spec.tsx
