+++ src/shared-components/EventListCard/Modal/Delete/EventListCardDeleteModal.tsx
