+++ src/shared-components/EventListCard/Modal/Preview/EventListCardPreviewModal.spec.tsx
