+++ src/shared-components/EventListCard/Modal/Preview/EventListCardPreviewModal.tsx
