+++ src/shared-components/EventListCard/Modal/updateLogic.ts
